chore: Abstract everything from top-level page/ components#236
