+++ components/About.styles.ts
@@ -1,0 +1,35 @@
+const styles = {
+	about: {
+		py: 10,
+		backgroundColor: '#fafafa',
+	},
+	sectionHeading: {
+		mb: 5,
+		fontSize: '3rem',
+		'@media (min-width: 600px)': {
+			fontSize: '4rem',
+		},
+	},
+	btn: {
+		fontWeight: 800,
+		fontStyle: 'italic',
+		fontSize: '1.2rem',
+		letterSpacing: '.5px',
+		color: '#fff',
+		mt: 5,
+		'@media (min-width: 600px)': {
+			fontSize: '1.6rem',
+		},
+	},
+	blurbBold: {
+		fontWeight: 600,
+		fontSize: '1.3rem',
+		mb: 0.5,
+	},
+	blurb: {
+		mb: 5,
+		fontSize: '1.3rem',
+	},
+}
+
+export default styles
