+++ components/Faq.styles.ts
@@ -1,0 +1,14 @@
+const styles = {
+	faq: {
+		py: 10,
+	},
+	faqHeading: {
+		mb: 5,
+		fontSize: '2.5rem',
+		'@media (min-width: 600px)': {
+			fontSize: '3.5rem',
+		},
+		textAlign: 'center',
+	},
+}
+export default styles
