+++ components/RecentProjectActivity.styles.ts
@@ -5,6 +5,7 @@ const styles = {
 	header: {
 		mb: 5,
 		fontSize: '3rem',
+		textAlign: 'center',
 		'@media (min-width: 600px)': {
 			fontSize: '4rem',
 		},
@@ -12,6 +13,7 @@ const styles = {
 	subheader: {
 		fontSize: '1.5rem',
 		mb: 7,
+		textAlign: 'center',
 		'@media (min-width: 600px)': {
 			fontSize: '1.75rem',
 			mb: 10,
