+++ components/RecentProjectActivity.tsx
@@ -1,4 +1,4 @@
-import { Box, Grid, Typography } from '@mui/material'
+import { Box, Container, Grid, Typography } from '@mui/material'
 import React from 'react'
 
 import ProjectCard from './ProjectCard'
@@ -11,17 +11,19 @@ type RecentProjectActivityProps = {
 const RecentProjectActivity = ({ projects }: RecentProjectActivityProps): JSX.Element => {
 	return (
 		<Box sx={styles.recentActivity} component="section">
-			<Typography variant="h2" sx={styles.header}>
-				Recent Activity
-			</Typography>
-			<Typography sx={styles.subheader}>Check out a few of the most recently updated projects on Arbor.</Typography>
-			<Grid container spacing={3}>
-				{projects.map(project => (
-					<Grid item xs={12} sm={4} key={project._id}>
-						<ProjectCard details={project} />
-					</Grid>
-				))}
-			</Grid>
+			<Container maxWidth="xl">
+				<Typography variant="h2" sx={styles.header}>
+					Recent Activity
+				</Typography>
+				<Typography sx={styles.subheader}>Check out a few of the most recently updated projects on Arbor.</Typography>
+				<Grid container spacing={3}>
+					{projects.map(project => (
+						<Grid item xs={12} sm={4} key={project._id}>
+							<ProjectCard details={project} />
+						</Grid>
+					))}
+				</Grid>
+			</Container>
 		</Box>
 	)
 }
